feat(SocialSharing): replace EnhancedSocialSharing with SocialSharing in dialog and update sharing functionality
